WIP - Preparation for log v3 format
In LogV2 the event records (prepare records) contain the stream name as a string. In LogV3 the stream will be identified by a long.
LogV3 also has a different record header format, which does not include the logical position of each record.

We'd prefer to keep using the existing Index and TF machinery for now since it is mature. This PR generalises that machinery to support
both log formats and their different ways of identifying streams.

TODO: There are some areas to be neatened up still, maked "//qq". Mostly around naming and locations now.